feat: Create New Activity & XML
- MainActivity : Display the list of added contacts
- ViewContact : Show the detailed info of the selected contact
- ResiterContact : Register new contact
