+++ app/src/main/AndroidManifest.xml
@@ -13,9 +13,15 @@
         android:theme="@style/Theme.Contacts"
         tools:targetApi="31">
         <activity
+            android:name=".ViewContact"
+            android:exported="false" />
+        <activity
+            android:name=".ResiterContact"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
-            android:windowSoftInputMode="adjustResize"
-            android:exported="true">
+            android:exported="true"
+            android:windowSoftInputMode="adjustResize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
