+++ app/src/main/res/layout/activity_view_contact.xml
@@ -5,8 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-
+    tools:context=".ViewContact">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
